Clang-tidy fixes

+++ packages/seacas/libraries/exodus/src/ex_create.c
@@ -330,7 +330,7 @@ int ex_create_int(const char *path, int cmode, int *comp_ws, int *io_ws, int run
 #endif
   }
   else {
-    filesiz = (int)((my_mode & EX_64BIT_OFFSET) || (ex_large_model(-1) == 1));
+    filesiz = ((my_mode & EX_64BIT_OFFSET) || (ex_large_model(-1) == 1));
   }
 
   if (
@@ -455,7 +455,7 @@ int ex_create_int(const char *path, int cmode, int *comp_ws, int *io_ws, int run
   }
 
   /* store Exodus file float word size  as an attribute */
-  lio_ws = (int)(*io_ws);
+  lio_ws = (*io_ws);
   if ((status = nc_put_att_int(exoid, NC_GLOBAL, ATT_FLT_WORDSIZE, NC_INT, 1, &lio_ws)) !=
       NC_NOERR) {
     snprintf(errmsg, MAX_ERR_LENGTH, "ERROR: failed to store Exodus II file float word size "
